+++ mleap-core/src/main/scala/ml/combust/mleap/core/feature/VectorIndexerModel.scala
@@ -1,5 +1,7 @@
 package ml.combust.mleap.core.feature
 
+import java.util.NoSuchElementException
+
 import ml.combust.mleap.core.Model
 import ml.combust.mleap.core.annotation.SparkCode
 import ml.combust.mleap.core.types.{StructType, TensorType}
@@ -8,12 +10,14 @@ import org.apache.spark.ml.linalg.{DenseVector, SparseVector, Vector}
 /**
   * Created by hollinwilkins on 12/28/16.
   */
-@SparkCode(uri = "https://github.com/apache/spark/blob/v2.0.0/mllib/src/main/scala/org/apache/spark/ml/feature/VectorIndexer.scala")
+@SparkCode(uri = "https://github.com/apache/spark/blob/v2.4.5/mllib/src/main/scala/org/apache/spark/ml/feature/VectorIndexer.scala")
 case class VectorIndexerModel(numFeatures: Int,
-                              categoryMaps: Map[Int, Map[Double, Int]]) extends Model {
+                              categoryMaps: Map[Int, Map[Double, Int]],
+                              handleInvalid: HandleInvalid = HandleInvalid.Error) extends Model {
   val sortedCatFeatureIndices = categoryMaps.keys.toArray.sorted
   val localVectorMap = categoryMaps
   val localNumFeatures = numFeatures
+  val localHandleInvalid = handleInvalid
 
   def apply(features: Vector): Vector = predict(features)
   def predict(features: Vector): Vector = {
@@ -21,16 +25,32 @@ case class VectorIndexerModel(numFeatures: Int,
       s" $numFeatures but found length ${features.size}")
     features match {
       case dv: DenseVector =>
+        var hasInvalid = false
         val tmpv = dv.copy
         localVectorMap.foreach { case (featureIndex: Int, categoryMap: Map[Double, Int]) =>
-          tmpv.values(featureIndex) = categoryMap(tmpv(featureIndex))
+          try {
+            tmpv.values(featureIndex) = categoryMap(tmpv(featureIndex))
+          } catch {
+            case _: NoSuchElementException =>
+              localHandleInvalid match {
+                case HandleInvalid.Error =>
+                  throw new IllegalArgumentException(s"VectorIndexer encountered invalid value " +
+                    s"${tmpv(featureIndex)} on feature index $featureIndex. To handle " +
+                    s"or skip invalid value, try setting VectorIndexer.handleInvalid.")
+                case HandleInvalid.Keep =>
+                  tmpv.values(featureIndex) = categoryMap.size
+                case HandleInvalid.Skip =>
+                  hasInvalid = true
+              }
+          }
         }
-        tmpv
+        if (hasInvalid) null else tmpv
       case sv: SparseVector =>
         // We use the fact that categorical value 0 is always mapped to index 0.
+        var hasInvalid = false
         val tmpv = sv.copy
         var catFeatureIdx = 0 // index into sortedCatFeatureIndices
-      var k = 0 // index into non-zero elements of sparse vector
+        var k = 0 // index into non-zero elements of sparse vector
         while (catFeatureIdx < sortedCatFeatureIndices.length && k < tmpv.indices.length) {
           val featureIndex = sortedCatFeatureIndices(catFeatureIdx)
           if (featureIndex < tmpv.indices(k)) {
@@ -38,12 +58,26 @@ case class VectorIndexerModel(numFeatures: Int,
           } else if (featureIndex > tmpv.indices(k)) {
             k += 1
           } else {
-            tmpv.values(k) = localVectorMap(featureIndex)(tmpv.values(k))
+            try {
+              tmpv.values(k) = localVectorMap(featureIndex)(tmpv.values(k))
+            } catch {
+              case _: NoSuchElementException =>
+                localHandleInvalid match {
+                  case HandleInvalid.Error =>
+                    throw new IllegalArgumentException(s"VectorIndexer encountered invalid value " +
+                      s"${tmpv.values(k)} on feature index $featureIndex. To handle " +
+                      s"or skip invalid value, try setting VectorIndexer.handleInvalid.")
+                  case HandleInvalid.Keep =>
+                    tmpv.values(k) = localVectorMap(featureIndex).size
+                  case HandleInvalid.Skip =>
+                    hasInvalid = true
+                }
+            }
             catFeatureIdx += 1
             k += 1
           }
         }
-        tmpv
+        if (hasInvalid) null else tmpv
     }
   }