+++ mleap-xgboost-runtime/src/test/scala/ml/combust/mleap/xgboost/runtime/testing/BundleSerializationUtils.scala
@@ -1,12 +1,13 @@
 package ml.combust.mleap.xgboost.runtime.testing
 
 import java.io.File
+import java.nio.file.{Files, Path}
 
 import ml.combust.bundle.BundleFile
 import ml.combust.bundle.serializer.SerializationFormat
 import ml.combust.mleap.runtime.{MleapContext, frame}
 import ml.combust.mleap.runtime.frame.Transformer
-import ml.combust.mleap.xgboost.runtime.bundle.ops.{XGBoostClassificationOp, XGBoostPredictorClassificationOp}
+import ml.combust.mleap.xgboost.runtime.bundle.ops.{XGBoostClassificationOp, XGBoostPredictorClassificationOp, XGBoostPredictorRegressionOp, XGBoostRegressionOp}
 import resource.managed
 
 
@@ -15,9 +16,13 @@ trait BundleSerializationUtils {
   def serializeModelToMleapBundle(transformer: Transformer): File = {
     import ml.combust.mleap.runtime.MleapSupport._
 
-    new File("/tmp/mleap/xgboost-runtime-parity").mkdirs()
-    val file = new File(s"/tmp/mleap/xgboost-runtime-parity/${this.getClass.getName}.zip")
-    file.delete()
+    val tempDirPath = {
+      val temp: Path = Files.createTempDirectory("xgboost-runtime-parity")
+      temp.toFile.deleteOnExit()
+      temp.toAbsolutePath
+    }
+
+    val file = new File(s"${tempDirPath}/${this.getClass.getName}.zip")
 
     for(bf <- managed(BundleFile(file))) {
       transformer.writeBundle.format(SerializationFormat.Json).save(bf).get
@@ -41,8 +46,10 @@ trait BundleSerializationUtils {
     // Register a different Op to change the deserialization class between tests.
     // Use to deserialize with Predictor rather than xgboost4j
     context.bundleRegistry.register(new XGBoostPredictorClassificationOp())
+    context.bundleRegistry.register(new XGBoostPredictorRegressionOp())
     val transformer = loadMleapTransformerFromBundle(bundleFile)
     context.bundleRegistry.register(new XGBoostClassificationOp())  // revert to the original Op
+    context.bundleRegistry.register(new XGBoostRegressionOp())  // revert to the original Op
     transformer
   }
 }