+++ mleap-xgboost-spark/src/main/scala/ml/dmlc/xgboost4j/scala/spark/mleap/XGBoostClassificationModelOp.scala
@@ -32,12 +32,15 @@ class XGBoostClassificationModelOp extends SimpleSparkOp[XGBoostClassificationMo
 
       val out = Files.newOutputStream(context.file("xgboost.model"))
       obj._booster.saveModel(out)
-
       val numFeatures = context.context.dataset.get.select(obj.getFeaturesCol).first.getAs[Vector](0).size
       model.withValue("thresholds", thresholds.map(_.toSeq).map(Value.doubleList)).
         withValue("num_classes", Value.int(obj.numClasses)).
         withValue("num_features", Value.int(numFeatures)).
-        withValue("tree_limit", Value.int(obj.getOrDefault(obj.treeLimit)))
+        withValue("tree_limit", Value.int(obj.getOrDefault(obj.treeLimit))).
+        withValue("missing", Value.float(obj.getOrDefault(obj.missing))).
+        withValue("infer_batch_size", Value.int(obj.getOrDefault(obj.inferBatchSize))).
+        withValue("use_external_memory", Value.boolean(obj.getOrDefault(obj.useExternalMemory))).
+        withValue("allow_non_zero_for_missing", Value.boolean(obj.getOrDefault(obj.allowNonZeroForMissing)))
     }
 
     override def load(model: Model)
@@ -46,14 +49,29 @@ class XGBoostClassificationModelOp extends SimpleSparkOp[XGBoostClassificationMo
         SXGBoost.loadModel(in)
       }).tried.get
 
-      new XGBoostClassificationModel("", model.value("num_classes").getInt, booster)
+      val xgb = new XGBoostClassificationModel("", model.value("num_classes").getInt, booster).
+          setTreeLimit(model.value("tree_limit").getInt)
+
+      model.getValue("thresholds").map(o => xgb.setThresholds(o.getDoubleList.toArray))
+      model.getValue("missing").map(o => xgb.setMissing(o.getFloat))
+      model.getValue("allow_non_zero_for_missing").map(o => xgb.setAllowZeroForMissingValue(o.getBoolean))
+      model.getValue("infer_batch_size").map(o => xgb.setInferBatchSize(o.getInt))
+      model.getValue("use_external_memory").map(o => xgb.set(xgb.useExternalMemory, o.getBoolean))
+      xgb
     }
   }
 
   override def sparkLoad(uid: String,
                          shape: NodeShape,
                          model: XGBoostClassificationModel): XGBoostClassificationModel = {
-    new XGBoostClassificationModel(uid, model.numClasses, model._booster)
+    val xgb = new XGBoostClassificationModel(uid, model.numClasses, model._booster)
+    if(model.isSet(model.thresholds)) xgb.setThresholds(model.getOrDefault(model.thresholds))
+    if(model.isSet(model.missing)) xgb.setMissing(model.getOrDefault(model.missing))
+    if(model.isSet(model.allowNonZeroForMissing)) xgb.setAllowZeroForMissingValue(model.getOrDefault(model.allowNonZeroForMissing))
+    if(model.isSet(model.inferBatchSize)) xgb.setInferBatchSize(model.getOrDefault(model.inferBatchSize))
+    if(model.isSet(model.treeLimit)) xgb.setTreeLimit(model.getOrDefault(model.treeLimit))
+    if(model.isSet(model.useExternalMemory)) xgb.set(xgb.useExternalMemory, model.getOrDefault(model.useExternalMemory))
+    xgb
   }
 
   override def sparkInputs(obj: XGBoostClassificationModel): Seq[ParamSpec] = {